+++ CHANGELOG.md
@@ -1142,7 +1142,7 @@
 
 - set `process.env.NODE_ENV` when running tasks with explicit `dev` values
   ([#136](https://github.com/feltjs/gro/pull/136))
-- remove `import.meta.env.DEV` support for now despite SvelteKit alignment
+- remove `import.meta.-env.DEV` support for now despite SvelteKit alignment
   ([#136](https://github.com/feltjs/gro/pull/136))
 
 ## 0.9.4